+++ src/tool/Util.js
@@ -26,11 +26,11 @@
 
     function loadError(){
         var name = requirements[this.url].name;
-        error(name,'can not fetch remote data of : '+this.url);
+        error(name,XERROR.COMPONENT.LOADERROR,'can not fetch remote data of : '+this.url);
     }
     function loadTimeout(){
         var name = requirements[this.url].name;
-        error(name,'load timeout : '+this.url);
+        error(name,XERROR.COMPONENT.LOADTIMEOUT,'load timeout : '+this.url);
     }
     function onload(){
         if(this.status===0 || //native
@@ -42,10 +42,8 @@
 
             txt.match(/<\s*template[^<>]*>([\s\S]*)<\s*\/\s*template\s*>/img)[0];
             var tmpl = RegExp.$1;
-            if(!tmpl){
-                error(name,'can not find tag <template> in component file');
-                return;
-            }
+            
+            assert(!tmpl,name,XERROR.COMPONENT.TEMPLATETAG,'can not find tag <template> in component file');
 
             var css = '';
             tmpl = tmpl.replace(/<\s*style[^<>]*>([\s\S]*?)<\s*\/\s*style\s*>/img,function(a,b){